@@ -1,82 +1,122 @@
-﻿@{
+﻿@model ESFA.DC.LARS.Web.Models.ViewModels.HomeViewModel
+@{
     ViewData["Title"] = "Home Page";
 }
 
-<div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds">
-        <div class="app-masthead">
-            <span class="govuk-heading-l white-text">Search for a qualification</span>
-            <p class="govuk-body white-text">Search by <strong>name, reference number</strong> or <strong>awarding body</strong></p>
-            <form id="changeUrl" class="govuk-form-group" action="all-results-func">
-                <div id="tt-overlay" class="autocomplete-wrapper">
-                    <div class="autocomplete__wrapper" role="combobox" aria-expanded="false">
-                        <div aria-atomic="true" aria-live="polite" role="status" style="border: 0px; clip: rect(0px, 0px, 0px, 0px); height: 1px; margin-bottom: -1px; margin-right: -1px; overflow: hidden; padding: 0px; position: absolute; white-space: nowrap; width: 1px;">
-                            No search results.<span>,,</span>
-                        </div>
-                        <input aria-owns="autocomplete-overlay__listbox" autocomplete="off" class="autocomplete__input autocomplete__input--default" id="autocomplete-overlay" name="input-autocomplete" placeholder="e.g maths or 40010740" type="text" role="textbox">
-                        <ul class="autocomplete__menu autocomplete__menu--overlay autocomplete__menu--hidden" id="autocomplete-overlay__listbox" role="listbox"></ul>
-                    </div>
-                </div>
-                <button id="searchProviders" type="search" class="govuk-button">
-                    <img src="https://assets.publishing.service.gov.uk/frontend/govuk_publishing_components/search-button-ca89b2a79f944909ceb7370d3f0b78811d32b96e883348fcd8886f63dd619585.png" />
-                </button>
-            </form>
-
-            <div class="flex space-between">
-                <div class="govuk-form-group govuk-!-margin-top-3 govuk-!-margin-bottom-1">
-                    <label class="govuk-label white-text" for="width-10">
-                        Awarding body
-                    </label>
-                    <input class="govuk-input govuk-input--width-10" id="width-10" name="width-10" type="text">
-
-                </div>
-
-                <div class="govuk-form-group govuk-!-margin-top-3 govuk-!-margin-bottom-1">
-                    <label class="govuk-label white-text" for="sort">
-                        Level
-                    </label>
-                    <select class="govuk-select" id="sort" name="sort">
-
-                        <option value="updated" selected>Level 1</option>
-                        <option value="updated">Level 1/2</option>
-                        <option value="updated">Level 2</option>
-                        <option value="updated">Level 3</option>
-                        <option value="updated">Level 4</option>
-                        <option value="updated">Level 5</option>
-                        <option value="updated">Level 6</option>
-                        <option value="updated">Level 7</option>
-                        <option value="updated">Level 8</option>
-                        <option value="updated">Entry level</option>
-                        <option value="updated">Higher level</option>
-                        <option value="updated">Mixed level</option>
-                        <option value="updated">Not known</option>
-                        <option value="updated">Level 2</option>
-
-
-                    </select>
-                </div>
-
-                <div class="govuk-form-group govuk-!-margin-top-3 govuk-!-margin-bottom-1">
-                    <label class="govuk-label white-text" for="sort">
-                        Teaching year
-                    </label>
-                    <select class="govuk-select" id="sort" name="sort">
-
-                        <option value="updated" selected>View all</option>
-                        <option value="updated">Valid for current year</option>
-                        <option value="updated">All 2019 to 2020</option>
-                        <option value="updated">2018 to 2019</option>
-                        <option value="updated">2017 to 2018</option>
-                        <option value="updated">2016 to 2017</option>
-
-
-
-                    </select>
-                </div>
+@section PageHeader
+{
+    <div class="app-masthead">
+        <div class="govuk-width-container">
+            <div class="govuk-grid-column-two-thirds no-padding">
+                <h1 class="govuk-heading-xl white-text">
+                    Find a learning aim
+                </h1>
+                <p class="govuk-body-l white-text">Find the latest information on available Qualifications, Standards, Apprenticeships, T Levels and Units.</p>
+                <p class="govuk-body white-text">Standards will show you information on funding, dates and common components. Qualifications and Units show you funding streams for courses and the last date learners can start. Apprenticeships and T Levels show you pathways, dates, and common components.</p>
             </div>
-
         </div>
     </div>
+}
+
+<div class="govuk-grid-row" id="homeApp">
+    <div class="govuk-grid-column-two-thirds">
+        <form asp-action="Index" asp-controller="Home" class="govuk-form-group" id="searchForm">
+            <label class="govuk-label">Search by <strong>learning aim</strong> or <strong>reference number</strong></label>
+                <span class="govuk-hint">
+                    For wildcard searches use the * symbol
+                </span>
+                <div id="tt-overlay" class="autocomplete-wrapper" aria-label="Search by name or reference number">
+                    <div class="autocomplete__wrapper" role="combobox" aria-expanded="false">
+                        <div aria-atomic="true" aria-live="polite" role="status" style="border: 0; clip: rect(0, 0, 0, 0); height: 1px; margin-bottom: -1px; margin-right: -1px; overflow: hidden; padding: 0; position: absolute; white-space: nowrap; width: 1px;">
+                            No search results.<span>,,</span>
+                        </div>
+                        <input aria-owns="autocomplete-overlay__listbox" autocomplete="off" class="govuk-input" id="autocomplete-overlay" name="searchTerm" placeholder="" type="text" role="textbox">
+                        <ul class="autocomplete__menu autocomplete__menu--overlay autocomplete__menu--hidden" id="autocomplete-overlay__listbox" role="listbox">
+                        </ul>
+                    </div>
+                </div>
+
+                <br />
+
+                <div class="flex space-between">
+                    <div id="typeOfLearning" class="govuk-form-group flex-1">
+                        <label class="govuk-label" for="searchType">
+                            Type of learning
+                        </label>
+                        <select class="govuk-select full-width" id="searchType" name="searchType" asp-for="@Model.SearchType" asp-items="Html.GetEnumSelectList<LearningType>()" onchange="learningTypeChanged();">
+                        </select>
+                    </div>
+                    <div class="govuk-form-group flex-1" id="academicYear">
+                        <label class="govuk-label" for="TeachingYear">
+                            Academic Year
+                        </label>
+                        @if (Model?.Lookups?.AcademicYearLookups != null)
+                        {
+                            <select class="govuk-select full-width" name="TeachingYear" id="TeachingYear">
+                                @foreach (var year in Model.Lookups.AcademicYearLookups)
+                                {
+                                    @if (year.IsCurrentAcademicYear)
+                                    {
+                                        <option value="@year.AcademicYear" selected>@year.AcademicYearDesc</option>
+                                    }
+                                    else
+                                    {
+                                        <option value="@year.AcademicYear">@year.AcademicYearDesc</option>
+                                    }
+                                }
+                            </select>
+                        }
+                    </div>
+                </div>
+
+            <br/>
+            <button id="search" type="search" class="govuk-button govuk-button--start govuk-!-margin-bottom-0" v-on:click="updateFilters()">
+                Search
+                <svg class="govuk-button__start-icon" xmlns="http://www.w3.org/2000/svg" width="17.5" height="19" viewBox="0 0 33 40" role="presentation" focusable="false">
+                    <path fill="currentColor" d="M0 0h13l20 20-20 20H0l20-20z"></path>
+                </svg>
+            </button>
+        </form>
+        </div>
+    
     <div class="govuk-grid-column-one-third">
+        <div class="white-side-bar">
+            <h2 class="govuk-heading-s" id="subsection-title">
+                Database downloads
+            </h2>
+            <nav role="navigation" aria-labelledby="subsection-title">
+                <ul class="govuk-list govuk-!-font-size-16">
+                    <li>
+                        <a class="govuk-link govuk-link govuk-link--no-visited-state" asp-controller="DownloadData" asp-action="Index">
+                            View current downloads
+                        </a>
+                    </li>
+                </ul>
+            </nav>
+        </div>
+    </div>
+    <div class="govuk-grid-column-one-third">&nbsp;</div>
+</div>
+
+<div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+        <validation-errors errors="@Model.ValidationErrors"></validation-errors>
     </div>
 </div>
+
+@section PageScripts
+{
+    <script src="/assets/dist/js/search.js" asp-append-version="true"></script>
+    <script>
+        function learningTypeChanged() {
+            var selected = document.getElementById('searchType').value;
+            var academicYear = document.getElementById('academicYear');
+            if (selected === "0" || selected === "2") {
+                // Hide the academic year dropdown the the learning type is not Qualification or Unit
+                academicYear.style.visibility = 'visible';
+            } else {
+                academicYear.style.visibility = 'hidden';
+            }
+        }
+    </script>
+}
