@@ -1,7 +1,7 @@
 ﻿<div class="govuk-phase-banner">
     <p class="govuk-phase-banner__content">
         <strong class="govuk-tag govuk-phase-banner__content__tag">BETA</strong>
-        <span class="govuk-phase-banner__text">This is a new service – your <a href="https://www.smartsurvey.co.uk/s/dct-beta-feedback/" target="_blank">feedback</a> will help us to improve it.</span>
+        <span class="govuk-phase-banner__text">This is a new service – your <a href="https://www.smartsurvey.co.uk/s/findalearningaimfeedbackv2/" target="_blank">feedback</a> will help us to continue to improve the find a learning aim service.</span>
     </p>
 </div>
 @if (ViewBag.PageId != "home-page" && ViewData.ContainsKey("BackUrl"))
