@@ -36,7 +36,7 @@
         </div>
 
         <div class="govuk-header__content">
-            <a href="/" id="proposition-name" class="govuk-header__link govuk-header__link--service-name">Learning Aim Reference Service</a>
+            <a href="/" id="proposition-name" class="govuk-header__link govuk-header__link--service-name">Find a Learning Aim</a>
         </div>
     </div>
 
