@@ -1,4 +1,7 @@
-﻿<footer class="govuk-footer " role="contentinfo">
+﻿
+@inject IAppVersionService AppVer
+
+<footer class="govuk-footer " role="contentinfo">
     <div class="govuk-width-container ">
 
         <div class="govuk-footer__meta">
